Fixup bad Last string position index logic. Make sure to handle the case where a single row takes us over a limit; we should try anyway. Made MaxParameters and MaxSqlLength virtual properties so we can set it per-provider bulk copy. Set for Postgres, oracle, sqlite, sqlserver

+++ Source/LinqToDB/DataProvider/BasicBulkCopy.cs
@@ -15,6 +15,9 @@ namespace LinqToDB.DataProvider
 
 	public class BasicBulkCopy
 	{
+		protected virtual int MaxParameters => 999;
+		protected virtual int MaxSqlLength => 100000;
+
 		public virtual BulkCopyRowsCopied BulkCopy<T>(BulkCopyType bulkCopyType, ITable<T> table, BulkCopyOptions options, IEnumerable<T> source)
 			where T : notnull
 		{
@@ -344,8 +347,8 @@ namespace LinqToDB.DataProvider
 			Action<MultipleRowsHelper>                prepFunction,
 			Action<MultipleRowsHelper,object,string?> addFunction,
 			Action<MultipleRowsHelper>                finishFunction,
-			int                                       maxParameters = 999,
-			int                                       maxSqlLength  = 100000)
+			int                                       maxParameters,
+			int                                       maxSqlLength)
 		{
 			prepFunction(helper);
 
@@ -356,17 +359,19 @@ namespace LinqToDB.DataProvider
 				addFunction(helper, item!, from);
 				var needRemove = helper.Parameters.Count > maxParameters ||
 				                 helper.StringBuilder.Length > maxSqlLength;
+				var isSingle = helper.CurrentCount == 1;
 				if (helper.CurrentCount >= helper.BatchSize || needRemove)
 				{
-					if (needRemove)
+					if (needRemove && !isSingle)
 					{
 						helper.Parameters.RemoveRange(helper.LastRowParameterIndex, helper.ParameterIndex-helper.LastRowParameterIndex);
-						helper.StringBuilder.Length = helper.LastRowStringIndex;	
+						helper.StringBuilder.Length = helper.LastRowStringIndex;
+						helper.RowsCopied.RowsCopied--;	
 					}
 					finishFunction(helper);
 					if (!helper.Execute())
 						return helper.RowsCopied;
-					if (needRemove)
+					if (needRemove && !isSingle)
 					{
 						addFunction(helper, item!, from);	
 					}
@@ -390,30 +395,32 @@ namespace LinqToDB.DataProvider
 			Action<MultipleRowsHelper, object, string?> addFunction,
 			Action<MultipleRowsHelper>                  finishFunction,
 			CancellationToken                           cancellationToken,
-			int                                         maxParameters = 999,
-			int                                         maxSqlLength  = 100000)
+			int                                         maxParameters,
+			int                                         maxSqlLength)
 		{
 			prepFunction(helper);
 
 			foreach (var item in source)
 			{
 				helper.LastRowParameterIndex = helper.ParameterIndex;
-				helper.LastRowStringIndex    = helper.LastRowStringIndex;
+				helper.LastRowStringIndex    = helper.StringBuilder.Length;
 				addFunction(helper, item!, from);
 
 				var needRemove = helper.Parameters.Count     > maxParameters ||
 				                 helper.StringBuilder.Length > maxSqlLength;
+				var isSingle = helper.CurrentCount == 1;
 				if (helper.CurrentCount >= helper.BatchSize || needRemove)
 				{
-					if (needRemove)
+					if (needRemove && !isSingle)
 					{
 						helper.Parameters.RemoveRange(helper.LastRowParameterIndex, helper.ParameterIndex-helper.LastRowParameterIndex);
-						helper.StringBuilder.Length = helper.LastRowStringIndex;	
+						helper.StringBuilder.Length = helper.LastRowStringIndex;
+						helper.RowsCopied.RowsCopied--;
 					}
 					finishFunction(helper);
 					if (!await helper.ExecuteAsync(cancellationToken).ConfigureAwait(Common.Configuration.ContinueOnCapturedContext))
 						return helper.RowsCopied;
-					if (needRemove)
+					if (needRemove && !isSingle)
 					{
 						addFunction(helper, item!, from);	
 					}
@@ -438,30 +445,32 @@ namespace LinqToDB.DataProvider
 			Action<MultipleRowsHelper, object, string?> addFunction,
 			Action<MultipleRowsHelper>                  finishFunction,
 			CancellationToken                           cancellationToken,
-			int                                         maxParameters = 999,
-			int                                         maxSqlLength  = 100000)
+			int                                         maxParameters,
+			int                                         maxSqlLength)
 		{
 			prepFunction(helper);
 
 			await foreach (var item in source.ConfigureAwait(Common.Configuration.ContinueOnCapturedContext).WithCancellation(cancellationToken))
 			{
 				helper.LastRowParameterIndex = helper.ParameterIndex;
-				helper.LastRowStringIndex    = helper.LastRowStringIndex;
+				helper.LastRowStringIndex    = helper.StringBuilder.Length;
 				addFunction(helper, item!, from);
 
 				var needRemove = helper.Parameters.Count     > maxParameters ||
 				                 helper.StringBuilder.Length > maxSqlLength;
+				var isSingle = helper.CurrentCount == 1;
 				if (helper.CurrentCount >= helper.BatchSize || needRemove)
 				{
-					if (needRemove)
+					if (needRemove && !isSingle)
 					{
 						helper.Parameters.RemoveRange(helper.LastRowParameterIndex, helper.ParameterIndex-helper.LastRowParameterIndex);
-						helper.StringBuilder.Length = helper.LastRowStringIndex;	
+						helper.StringBuilder.Length = helper.LastRowStringIndex;
+						helper.RowsCopied.RowsCopied--;	
 					}
 					finishFunction(helper);
 					if (!await helper.ExecuteAsync(cancellationToken).ConfigureAwait(Common.Configuration.ContinueOnCapturedContext))
 						return helper.RowsCopied;
-					if (needRemove)
+					if (needRemove && !isSingle)
 					{
 						addFunction(helper, item!, from);	
 					}
@@ -483,14 +492,14 @@ namespace LinqToDB.DataProvider
 			=> MultipleRowsCopy1(new MultipleRowsHelper<T>(table, options), source);
 
 		protected BulkCopyRowsCopied MultipleRowsCopy1(MultipleRowsHelper helper, IEnumerable source)
-			=> MultipleRowsCopyHelper(helper, source, null, MultipleRowsCopy1Prep, MultipleRowsCopy1Add, MultipleRowsCopy1Finish);
+			=> MultipleRowsCopyHelper(helper, source, null, MultipleRowsCopy1Prep, MultipleRowsCopy1Add, MultipleRowsCopy1Finish,MaxParameters, MaxSqlLength);
 
 		protected Task<BulkCopyRowsCopied> MultipleRowsCopy1Async<T>(ITable<T> table, BulkCopyOptions options, IEnumerable<T> source, CancellationToken cancellationToken)
 			where T : notnull
 			=> MultipleRowsCopy1Async(new MultipleRowsHelper<T>(table, options), source, cancellationToken);
 
 		protected Task<BulkCopyRowsCopied> MultipleRowsCopy1Async(MultipleRowsHelper helper, IEnumerable source, CancellationToken cancellationToken)
-			=> MultipleRowsCopyHelperAsync(helper, source, null, MultipleRowsCopy1Prep, MultipleRowsCopy1Add, MultipleRowsCopy1Finish, cancellationToken);
+			=> MultipleRowsCopyHelperAsync(helper, source, null, MultipleRowsCopy1Prep, MultipleRowsCopy1Add, MultipleRowsCopy1Finish, cancellationToken, MaxParameters, MaxSqlLength);
 
 #if NATIVE_ASYNC
 		protected Task<BulkCopyRowsCopied> MultipleRowsCopy1Async<T>(ITable<T> table, BulkCopyOptions options, IAsyncEnumerable<T> source, CancellationToken cancellationToken)
@@ -499,7 +508,7 @@ namespace LinqToDB.DataProvider
 
 		protected Task<BulkCopyRowsCopied> MultipleRowsCopy1Async<T>(MultipleRowsHelper helper, IAsyncEnumerable<T> source, CancellationToken cancellationToken)
 		where T: notnull
-			=> MultipleRowsCopyHelperAsync(helper, source, null, MultipleRowsCopy1Prep, MultipleRowsCopy1Add, MultipleRowsCopy1Finish, cancellationToken);
+			=> MultipleRowsCopyHelperAsync(helper, source, null, MultipleRowsCopy1Prep, MultipleRowsCopy1Add, MultipleRowsCopy1Finish, cancellationToken, MaxParameters, MaxSqlLength);
 #endif
 
 		private void MultipleRowsCopy1Prep(MultipleRowsHelper helper)
@@ -551,14 +560,14 @@ namespace LinqToDB.DataProvider
 			=> MultipleRowsCopy2(new MultipleRowsHelper<T>(table, options), source, from);
 
 		protected BulkCopyRowsCopied MultipleRowsCopy2(MultipleRowsHelper helper, IEnumerable source, string from)
-			=> MultipleRowsCopyHelper(helper, source, from, MultipleRowsCopy2Prep, MultipleRowsCopy2Add, MultipleRowsCopy2Finish);
+			=> MultipleRowsCopyHelper(helper, source, from, MultipleRowsCopy2Prep, MultipleRowsCopy2Add, MultipleRowsCopy2Finish, MaxParameters, MaxSqlLength);
 
 		protected Task<BulkCopyRowsCopied> MultipleRowsCopy2Async<T>(ITable<T> table, BulkCopyOptions options, IEnumerable<T> source, string from, CancellationToken cancellationToken)
 			where T : notnull
 			=> MultipleRowsCopy2Async(new MultipleRowsHelper<T>(table, options), source, from, cancellationToken);
 
 		protected Task<BulkCopyRowsCopied> MultipleRowsCopy2Async(MultipleRowsHelper helper, IEnumerable source, string from, CancellationToken cancellationToken)
-			=> MultipleRowsCopyHelperAsync(helper, source, from, MultipleRowsCopy2Prep, MultipleRowsCopy2Add, MultipleRowsCopy2Finish, cancellationToken);
+			=> MultipleRowsCopyHelperAsync(helper, source, from, MultipleRowsCopy2Prep, MultipleRowsCopy2Add, MultipleRowsCopy2Finish, cancellationToken, MaxParameters, MaxSqlLength);
 
 #if NATIVE_ASYNC
 		protected Task<BulkCopyRowsCopied> MultipleRowsCopy2Async<T>(ITable<T> table, BulkCopyOptions options, IAsyncEnumerable<T> source, string from, CancellationToken cancellationToken)
@@ -567,7 +576,7 @@ namespace LinqToDB.DataProvider
 
 		protected Task<BulkCopyRowsCopied> MultipleRowsCopy2Async<T>(MultipleRowsHelper helper, IAsyncEnumerable<T> source, string from, CancellationToken cancellationToken)
 		where T: notnull
-			=> MultipleRowsCopyHelperAsync(helper, source, from, MultipleRowsCopy2Prep, MultipleRowsCopy2Add, MultipleRowsCopy2Finish, cancellationToken);
+			=> MultipleRowsCopyHelperAsync(helper, source, from, MultipleRowsCopy2Prep, MultipleRowsCopy2Add, MultipleRowsCopy2Finish, cancellationToken, MaxParameters, MaxSqlLength);
 #endif
 
 		private void MultipleRowsCopy2Prep(MultipleRowsHelper helper)
@@ -612,15 +621,15 @@ namespace LinqToDB.DataProvider
 		}
 
 		protected BulkCopyRowsCopied MultipleRowsCopy3(MultipleRowsHelper helper, BulkCopyOptions options, IEnumerable source, string from)
-			=> MultipleRowsCopyHelper(helper, source, from, MultipleRowsCopy3Prep, MultipleRowsCopy3Add, MultipleRowsCopy3Finish);
+			=> MultipleRowsCopyHelper(helper, source, from, MultipleRowsCopy3Prep, MultipleRowsCopy3Add, MultipleRowsCopy3Finish, MaxParameters, MaxSqlLength);
 
 		protected Task<BulkCopyRowsCopied> MultipleRowsCopy3Async(MultipleRowsHelper helper, BulkCopyOptions options, IEnumerable source, string from, CancellationToken cancellationToken)
-			=> MultipleRowsCopyHelperAsync(helper, source, from, MultipleRowsCopy3Prep, MultipleRowsCopy3Add, MultipleRowsCopy3Finish, cancellationToken);
+			=> MultipleRowsCopyHelperAsync(helper, source, from, MultipleRowsCopy3Prep, MultipleRowsCopy3Add, MultipleRowsCopy3Finish, cancellationToken, MaxParameters, MaxSqlLength);
 
 #if NATIVE_ASYNC
 		protected Task<BulkCopyRowsCopied> MultipleRowsCopy3Async<T>(MultipleRowsHelper helper, BulkCopyOptions options, IAsyncEnumerable<T> source, string from, CancellationToken cancellationToken)
 		where T: notnull
-			=> MultipleRowsCopyHelperAsync(helper, source, from, MultipleRowsCopy3Prep, MultipleRowsCopy3Add, MultipleRowsCopy3Finish, cancellationToken);
+			=> MultipleRowsCopyHelperAsync(helper, source, from, MultipleRowsCopy3Prep, MultipleRowsCopy3Add, MultipleRowsCopy3Finish, cancellationToken, MaxParameters, MaxSqlLength);
 #endif
 
 		private void MultipleRowsCopy3Prep(MultipleRowsHelper helper)
